requirements.txt
requests
bs4
pymongo
python-dotenv